requirements.txt
# Core data processing
netCDF4==1.6.4
pandas==2.1.1
numpy==1.24.3
xarray==2023.8.0

# Database connections
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# Vector databases
faiss-cpu==1.7.4
chromadb==0.4.15

# Machine Learning & AI
openai==0.28.1
langchain==0.0.309
sentence-transformers==2.2.2
transformers==4.34.0

# Web framework and visualization
streamlit==1.27.2
plotly==5.17.0
folium==0.14.0
dash==2.14.1
dash-leaflet==0.1.23

# Geospatial processing
geopandas==0.14.0
shapely==2.0.1
pyproj==3.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
pydantic==2.4.2
fastapi==0.103.2
uvicorn==0.23.2

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Additional ML utilities
scikit-learn==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
